Clean up argument handling
The arguments are now handled better.  First, all arguments except "order" have
a default value.  Specifically, fs defaults to 1.0, and max_edge_freq defaults
to half the sampling frequency.  Secondly, the order of arguments was changed to
make more logical sense.  And finally, the argument order in __gen_filter_bank()
has been changed to match the signature of the constructor.